warnings

+++ src/IW5/IW5.cpp
@@ -8,7 +8,7 @@
 namespace IW5
 {
 
-auto opcode_size(opcode id) -> std::size_t
+auto opcode_size(opcode id) -> std::uint32_t
 {
 	switch (id)
 	{
